users.s3Upload -> attachments.create
